chore: 升级至 Spring Boot 4.0 / Java 21，适配 Jackson 3 及依赖变更
- 升级 java.version 至 21，同步更新文档中的版本要求
- pom.xml: spring-boot-starter-aop → spring-boot-starter-aspectj
- 修复 Spring Boot 4.0 自动配置类包路径变化 (EntityScan, DataSourceAutoConfiguration, HibernatePropertiesCustomizer)
- JacksonConfig: Jackson2ObjectMapperBuilderCustomizer → JsonMapperBuilderCustomizer
- 自定义序列化/反序列化类迁移至 Jackson 3 (com.fasterxml.jackson → tools.jackson, ValueSerializer/ValueDeserializer, @JacksonComponent)
- PageJsonSerializer: 适配 Jackson 3 JsonGenerator API (write*Property)
- CachingJsonRequestBodyFilter: ContentCachingRequestWrapper 构造函数增加 buffer 参数
- MinioFileOperator: Minio 9.0 stream() 参数 int → long

+++ src/main/java/io/github/jiangood/openadmin/modules/system/file/MinioFileOperator.java
@@ -37,7 +37,7 @@ public class MinioFileOperator implements FileOperator {
         PutObjectArgs arg = PutObjectArgs.builder()
                 .bucket(bucketName)
                 .object(key)
-                .stream(inputStream, inputStream.available(), -1)
+                .stream(inputStream, (long) inputStream.available(), -1L)
                 .build();
 
         this.client.putObject(arg);
